Integration with Django's migration system
``django-pghistory`` upgrades ``django-pgtrigger``, which now integrates
with the Django migration system.

Run ``python manage.py makemigrations`` to make migrations for the triggers
created by ``django-pghistory`` in order to upgrade.

If you are tracking changes to third-party models, register the tracker on
a proxy model so that migrations are created in the proper app.

Type: api-break