Correct type hint of def track(obj_field=)
When I upgraded pghistory, PyCharm showed me a type error that None is not expected, when it  actually is one of the allowed types. I can surpress this inspection for the entire class, but that is overkill. A correct type hint would be preferable, since None is a valid value and necessary for M2M table add/remove tracking.